extend and reserve DDL implemented

+++ rCancel.php
@@ -4,13 +4,14 @@
     $isbn = $_POST['isbn'];
     $cno = $_SESSION['cno'];
 
-    // echo "$isbn";
-    // echo "$cno";
-    // $conn -> beginTransaction();
-    $stmt = $conn -> prepare('DELETE FROM RESERVE
+    if($isbn == ""){
+        echo "<script>alert('도서를 선택해주세요.'); history.back();</script>"; 
+    }else{
+        $stmt = $conn -> prepare('DELETE FROM RESERVE
                               WHERE ISBN = :isbn
                               AND CNO = :cno
-    ');
-    $stmt -> execute(array($isbn, $cno));
-    echo "<script>alert('예약이 취소되었습니다.'); history.back();</script>"; 
+        ');
+        $stmt -> execute(array($isbn, $cno));
+        echo "<script>alert('예약이 취소되었습니다.'); history.back();</script>"; 
+    }
 ?>
